refactor(ecommerce): 重构金额相关代码
- 移除 Amount 类和 AmountCastTransformer 类- 使用 Money 类替换金额相关的代码
- 更新数据库迁移文件，将 bigint 类型改为 decimal 类型- 调整相关服务和模型，使用新的 Money 类

+++ database/migrations/create_product_skus_table.php
@@ -18,9 +18,9 @@ return new class extends Migration {
 
             // SKU 信息
             $table->string('currency', 10)->default('CNY')->comment('货币');
-            $table->bigInteger('price')->default(0)->comment('销售价');
-            $table->bigInteger('market_price')->nullable()->comment('市场价');
-            $table->bigInteger('cost_price')->nullable()->comment('成本价');
+            $table->decimal('price',12)->default(0)->comment('销售价');
+            $table->decimal('market_price',12)->nullable()->comment('市场价');
+            $table->decimal('cost_price',12)->nullable()->comment('成本价');
             // 库存
             $table->bigInteger('stock')->default(0)->comment('库存');
             $table->bigInteger('channel_stock')->default(0)->comment('渠道库存');
